refactor: modularize MidiAppClient into core/feedback/index modules
- Split src/midi/appClient.ts (>150 lines) into smaller modules:
  - core.ts: helper functions, markAppOutgoingAndForward, passthrough checks
  - feedback.ts: MIDI input port management and feedback wiring
  - index.ts: main MidiAppClient class delegating to helpers
- Update imports in controlMidiSender, midiBridge, voicemeeter
- Maintain same public API and functionality
- All tests pass (78/78)
- Follows project guideline: max 100-150 lines per file
