+++ TASKS.md
@@ -27,10 +27,12 @@
 - [ ] Learn page (assistant): enchaîner plusieurs learns pour produire un bloc `controls`
 - [ ] Tests de perf/jitter (< 20 ms) et micro-bench sur hot paths
 - [ ] CI GitHub Actions: pnpm i --frozen-lockfile, lint, check:types, test
- - [ ] Refactor MIDI: extraire un client partagé `MidiAppClient` (ports IN/OUT, cache, feedback → Router, optimistic shadow) et l'utiliser dans `drivers/midiBridge.ts` et `services/controlMidiSender.ts` pour supprimer la duplication
- - [ ] Clarifier l'API: renommer `controlMidiSender` → `midiAppClient` (ou similaire) et adapter `Router.handleControl()` pour appeler ce client générique pour `controls.*.midi`
- - [ ] Garder `MidiBridgeDriver` focalisé sur le « passthrough » page-scopé (X‑Touch ⇄ App) en s'appuyant sur `MidiAppClient` pour l'accès port/émission; éviter d'y intégrer la logique `controls.midi`
- - [ ] Tests: couvrir les conversions PB→CC (14b→7b) et la résolution `channelForCc` via `resolvePbToCcMappingForApp()` côté `controlMidiSender`/client partagé
+- [x] Refactor MIDI: extraire un client partagé `MidiAppClient` (ports IN/OUT, cache, feedback → Router, optimistic shadow) et l'utiliser dans `drivers/midiBridge.ts` et `services/controlMidiSender.ts` pour supprimer la duplication
+- [ ] Clarifier l'API: renommer `controlMidiSender` → `midiAppClient` (ou similaire) et adapter `Router.handleControl()` pour appeler ce client générique pour `controls.*.midi`
+- [x] Garder `MidiBridgeDriver` focalisé sur le « passthrough » page-scopé (X‑Touch ⇄ App) en s'appuyant sur `MidiAppClient` pour l'accès port/émission; éviter d'y intégrer la logique `controls.midi`
+- [ ] Tests: couvrir les conversions PB→CC (14b→7b) et la résolution `channelForCc` via `resolvePbToCcMappingForApp()` côté `controlMidiSender`/client partagé
+- [x] Diviser `src/midi/appClient.ts` (>150 lignes) en modules plus petits: `midi/appClient/core.ts` (send/convert/ports), `midi/appClient/feedback.ts` (wiring IN), `midi/appClient/index.ts` (API publique)
+- [ ] JSDoc sur `midi/appClient/*` et points de contact service/driver mis à jour
 
 ## En cours
 - [x] Pages: support d'un bloc `pages_global` (defaults fusionnés dans chaque page; override par page)
